Update app screenshots
The existing screen shots have become a lot out-dated as the user
interface of the app has been updated a lot. So, replace them
with news ones.

The newly added screen shots are in compressed jpeg format. This
brings down the size of images alone from 1.23MiB to 176.83KiB.
This would result in the site being served faster than before due
to the less size taken for the screen shots.